move all excalidraw element conversions to transformToExcalidrawSkeleton
