+++ src/converter/types/sequence.ts
@@ -4,120 +4,12 @@ import { GraphConverter } from "../GraphConverter.js";
 import { Sequence } from "../../parser/sequence.js";
 import { nanoid } from "nanoid";
 import { ExcalidrawElement } from "../../types.js";
-import { Arrow, Line, Node, Text } from "../../elementSkeleton.js";
-
-export const createLine = (line: Line) => {
-  const lineElement: ExcalidrawElementSkeleton = {
-    type: "line",
-    x: line.startX,
-    y: line.startY,
-    points: [
-      [0, 0],
-      [line.endX - line.startX, line.endY - line.startY],
-    ],
-    width: line.endX - line.startX,
-    height: line.endY - line.startY,
-    strokeStyle: line.strokeStyle || "solid",
-    strokeColor: line.strokeColor || "#000",
-    strokeWidth: line.strokeWidth || 1,
-  };
-  if (line.groupId) {
-    Object.assign(lineElement, { groupIds: [line.groupId] });
-  }
-  if (line.id) {
-    Object.assign(lineElement, { id: line.id });
-  }
-  return lineElement;
-};
-
-export const createText = (element: Text) => {
-  const textElement: ExcalidrawElementSkeleton = {
-    type: "text",
-    x: element.x,
-    y: element.y,
-    width: element.width,
-    height: element.height,
-    text: element.text || "",
-    fontSize: element.fontSize,
-    verticalAlign: "middle",
-  };
-  if (element.groupId) {
-    Object.assign(textElement, { groupIds: [element.groupId] });
-  }
-  if (element.id) {
-    Object.assign(textElement, { id: element.id });
-  }
-  return textElement;
-};
-
-export const createContainer = (
-  element: Exclude<Node, Line | Arrow | Text>
-) => {
-  let extraProps = {};
-  if (element.type === "rectangle" && element.subtype === "activation") {
-    extraProps = {
-      backgroundColor: "#e9ecef",
-      fillStyle: "solid",
-    };
-  }
-  const container: ExcalidrawElementSkeleton = {
-    id: element.id,
-    type: element.type,
-    x: element.x,
-    y: element.y,
-    width: element.width,
-    height: element.height,
-    label: {
-      text: element?.label?.text || "",
-      fontSize: element?.label?.fontSize,
-      verticalAlign: element.label?.verticalAlign || "middle",
-      strokeColor: element.label?.color || "#000",
-      groupIds: element.groupId ? [element.groupId] : [],
-    },
-    strokeStyle: element?.strokeStyle,
-    strokeWidth: element?.strokeWidth,
-    strokeColor: element?.strokeColor,
-    backgroundColor: element?.bgColor,
-    fillStyle: "solid",
-    ...extraProps,
-  };
-  if (element.groupId) {
-    Object.assign(container, { groupIds: [element.groupId] });
-  }
-
-  return container;
-};
-
-export const createArrow = (arrow: Arrow) => {
-  const arrowElement: ExcalidrawElementSkeleton = {
-    type: "arrow",
-    x: arrow.startX,
-    y: arrow.startY,
-    points: arrow.points || [
-      [0, 0],
-      [arrow.endX - arrow.startX, arrow.endY - arrow.startY],
-    ],
-    width: arrow.endX - arrow.startX,
-    height: arrow.endY - arrow.startY,
-    strokeStyle: arrow?.strokeStyle || "solid",
-    endArrowhead: arrow?.endArrowhead || null,
-    startArrowhead: arrow?.startArrowhead || null,
-    label: {
-      text: arrow?.label?.text || "",
-      fontSize: 16,
-    },
-    roundness: {
-      type: 2,
-    },
-    start: arrow.start,
-    end: arrow.end,
-  };
-  if (arrow.groupId) {
-    Object.assign(arrowElement, { groupIds: [arrow.groupId] });
-  }
-
-  return arrowElement;
-};
+import {
+  transformToExcalidrawLineSkeleton,
+  transformToExcalidrawTextSkeleton,
+  transformToExcalidrawContainerSkeleton,
+  transformToExcalidrawArrowSkeleton,
+} from "../transformToExcalidrawSkeleton.js";
 
 export const SequenceToExcalidrawSkeletonConvertor = new GraphConverter({
   converter: (chart: Sequence) => {
@@ -132,16 +24,16 @@ export const SequenceToExcalidrawSkeletonConvertor = new GraphConverter({
 
         switch (element.type) {
           case "line":
-            excalidrawElement = createLine(element);
+            excalidrawElement = transformToExcalidrawLineSkeleton(element);
             break;
 
           case "rectangle":
           case "ellipse":
-            excalidrawElement = createContainer(element);
+            excalidrawElement = transformToExcalidrawContainerSkeleton(element);
             break;
 
           case "text":
-            excalidrawElement = createText(element);
+            excalidrawElement = transformToExcalidrawTextSkeleton(element);
             break;
           default:
             throw `unknown type ${element.type}`;
@@ -159,7 +51,7 @@ export const SequenceToExcalidrawSkeletonConvertor = new GraphConverter({
       if (!line) {
         return;
       }
-      elements.push(createLine(line));
+      elements.push(transformToExcalidrawLineSkeleton(line));
     });
 
     Object.values(chart.arrows).forEach((arrow) => {
@@ -167,9 +59,11 @@ export const SequenceToExcalidrawSkeletonConvertor = new GraphConverter({
         return;
       }
 
-      elements.push(createArrow(arrow));
+      elements.push(transformToExcalidrawArrowSkeleton(arrow));
       if (arrow.sequenceNumber) {
-        elements.push(createContainer(arrow.sequenceNumber));
+        elements.push(
+          transformToExcalidrawContainerSkeleton(arrow.sequenceNumber)
+        );
       }
     });
     elements.push(...activations);
@@ -178,13 +72,13 @@ export const SequenceToExcalidrawSkeletonConvertor = new GraphConverter({
     if (chart.loops) {
       const { lines, texts, nodes } = chart.loops;
       lines.forEach((line) => {
-        elements.push(createLine(line));
+        elements.push(transformToExcalidrawLineSkeleton(line));
       });
       texts.forEach((text) => {
-        elements.push(createText(text));
+        elements.push(transformToExcalidrawTextSkeleton(text));
       });
       nodes.forEach((node) => {
-        elements.push(createContainer(node));
+        elements.push(transformToExcalidrawContainerSkeleton(node));
       });
     }
 
@@ -226,7 +120,7 @@ export const SequenceToExcalidrawSkeletonConvertor = new GraphConverter({
         const groupRectWidth = maxX - minX + PADDING * 2;
         const groupRectHeight = maxY - minY + PADDING * 2;
         const groupRectId = nanoid();
-        const groupRect = createContainer({
+        const groupRect = transformToExcalidrawContainerSkeleton({
           type: "rectangle",
           x: groupRectX,
           y: groupRectY,
